[S06P22D105-110] 게시글 기능 - 회원 연결

See merge request s06-bigdata-rec-sub2/S06P22D105!81

+++ backend/src/main/java/com/ladder/perfumism/article/service/ArticleService.java
@@ -8,6 +8,8 @@ import com.ladder.perfumism.article.domain.ArticleRepository;
 import com.ladder.perfumism.article.domain.ArticleSubject;
 import com.ladder.perfumism.global.exception.BusinessException;
 import com.ladder.perfumism.global.exception.ErrorCode;
+import com.ladder.perfumism.member.domain.Member;
+import com.ladder.perfumism.member.domain.MemberRepository;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -18,15 +20,27 @@ import org.springframework.transaction.annotation.Transactional;
 public class ArticleService {
 
     private final ArticleRepository articleRepository;
+    private final MemberRepository memberRepository;
 
 
-    public ArticleService(ArticleRepository articleRepository){
+    public ArticleService(ArticleRepository articleRepository, MemberRepository memberRepository){
         this.articleRepository = articleRepository;
+        this.memberRepository = memberRepository;
+    }
+
+    private void checkArticleOwner(String email, Article article){
+        if(!article.getMember().getEmail().equals(email)){
+            throw new BusinessException(ErrorCode.ARTICLE_IS_NOT_YOURS);
+        }
     }
 
     @Transactional
-    public void articleCreate(ArticleCreateRequest request) {
+    public void articleCreate(String email, ArticleCreateRequest request) {
+        Member member = memberRepository.findByEmail(email)
+            .orElseThrow(()-> new BusinessException(ErrorCode.MEMBER_NOT_FOUND_BY_EMAIL));
+
         Article article = Article.builder()
+            .member(member)
             .title(request.getTitle())
             .content(request.getContent())
             .subject(request.getSubject())
@@ -36,7 +50,10 @@ public class ArticleService {
     }
 
     @Transactional
-    public ArticleReadListResponse showArticleList(Pageable pageable, ArticleSubject subject) {
+    public ArticleReadListResponse showArticleList(String email,Pageable pageable, ArticleSubject subject) {
+        Member member = memberRepository.findByEmail(email)
+            .orElseThrow(()->new BusinessException(ErrorCode.MEMBER_NOT_FOUND_BY_EMAIL));
+
         Page<Article> articleList;
         if (subject != null){
             articleList = articleRepository.findBySubject(subject,pageable);
@@ -49,16 +66,23 @@ public class ArticleService {
     }
 
     @Transactional
-    public ArticleReadDetailResponse showArticleDetail(Long article_Id) {
+    public ArticleReadDetailResponse showArticleDetail(String email, Long article_Id) {
+        Member member = memberRepository.findByEmail(email)
+            .orElseThrow(()->new BusinessException(ErrorCode.MEMBER_NOT_FOUND_BY_EMAIL));
         Article article = articleRepository.findById(article_Id)
-            .orElseThrow(()-> new BusinessException(ErrorCode.ARTICLE_NOT_FOUNT_MY_ARTICLE_ID));
+            .orElseThrow(()-> new BusinessException(ErrorCode.ARTICLE_NOT_FOUND));
         return ArticleReadDetailResponse.from(article);
     }
 
     @Transactional
-    public void updateArticle(Long articleId, ArticleCreateRequest request) {
+    public void updateArticle(String email, Long articleId, ArticleCreateRequest request) {
+        Member member = memberRepository.findByEmail(email)
+            .orElseThrow(()->new BusinessException(ErrorCode.MEMBER_NOT_FOUND_BY_EMAIL));
+
         Article article = articleRepository.findById(articleId)
-            .orElseThrow(() -> new BusinessException(ErrorCode.ARTICLE_NOT_FOUNT_MY_ARTICLE_ID));
+            .orElseThrow(() -> new BusinessException(ErrorCode.ARTICLE_NOT_FOUND));
+
+        checkArticleOwner(email, article);
 
         article.changeSubject(request.getSubject());
         article.changeTitle(request.getTitle());
@@ -67,9 +91,14 @@ public class ArticleService {
     }
 
     @Transactional
-    public void deleteArticle(Long articleId) {
+    public void deleteArticle(String email,Long articleId) {
+        Member member = memberRepository.findByEmail(email)
+            .orElseThrow(()->new BusinessException(ErrorCode.MEMBER_NOT_FOUND_BY_EMAIL));
+
         Article article = articleRepository.findById(articleId)
-            .orElseThrow(()-> new BusinessException(ErrorCode.ARTICLE_NOT_FOUNT_MY_ARTICLE_ID));
+            .orElseThrow(()-> new BusinessException(ErrorCode.ARTICLE_NOT_FOUND));
+
+        checkArticleOwner(email, article);
 
         articleRepository.delete(article);
     }
